Move vcharts index.js up one directory
