+++ src/vcharts/index.js
@@ -2,16 +2,16 @@ import d3 from 'd3';
 import vg from 'vega';
 
 let templates = {
-  axis: require('../templates/axis.json'),
-  bar: require('../templates/bar.json'),
-  box: require('../templates/box.json'),
-  bullet: require('../templates/bullet.json'),
-  gantt: require('../templates/gantt.json'),
-  histogram: require('../templates/histogram.json'),
-  scatter: require('../templates/scatter.json'),
-  vega: require('../templates/vega.json'),
-  xy: require('../templates/xy.json'),
-  xymatrix: require('../templates/xymatrix.json')
+  axis: require('./templates/axis.json'),
+  bar: require('./templates/bar.json'),
+  box: require('./templates/box.json'),
+  bullet: require('./templates/bullet.json'),
+  gantt: require('./templates/gantt.json'),
+  histogram: require('./templates/histogram.json'),
+  scatter: require('./templates/scatter.json'),
+  vega: require('./templates/vega.json'),
+  xy: require('./templates/xy.json'),
+  xymatrix: require('./templates/xymatrix.json')
 };
 
 let getNestedRec = function (spec, parts) {
